# platform-v2/packages/contracts/openapi/platform-api.v1.json
{
  "components": {
    "schemas": {
      "SourceEndpoint": {
        "properties": {
          "method": {
            "example": "GET",
            "type": "string"
          },
          "type": {
            "example": "WFS",
            "type": "string"
          },
          "url": {
            "format": "uri",
            "type": "string"
          }
        },
        "required": [
          "type",
          "url",
          "method"
        ],
        "type": "object"
      },
      "SourceRegistryEntry": {
        "properties": {
          "accessClass": {
            "enum": [
              "A",
              "B",
              "C",
              "D",
              "E",
              "F"
            ],
            "example": "B",
            "type": "string"
          },
          "authority": {
            "example": "Prefeitura de São Paulo / GeoSampa",
            "type": "string"
          },
          "cadence": {
            "example": "diária/semanal",
            "type": "string"
          },
          "datasetCode": {
            "example": "LOTE_FISCAL",
            "type": "string"
          },
          "endpoints": {
            "items": {
              "$ref": "#/components/schemas/SourceEndpoint"
            },
            "type": "array"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "ingestionStatus": {
            "enum": [
              "healthy",
              "degraded",
              "broken",
              "manual",
              "disabled"
            ],
            "type": "string"
          },
          "lastCheckedAt": {
            "format": "date-time",
            "type": "string"
          },
          "municipalityIbge": {
            "example": "3550308",
            "type": "string"
          },
          "municipalityName": {
            "example": "São Paulo",
            "type": "string"
          },
          "parserVersion": {
            "example": "geosampa-lote-v1",
            "type": "string"
          },
          "sourceCode": {
            "example": "PMSP_GEOSAMPA_LOTES",
            "type": "string"
          },
          "sourceType": {
            "example": "WFS",
            "type": "string"
          }
        },
        "required": [
          "id",
          "sourceCode",
          "sourceType",
          "authority",
          "datasetCode",
          "accessClass",
          "ingestionStatus",
          "endpoints"
        ],
        "type": "object"
      }
    }
  },
  "info": {
    "contact": {},
    "description": "Versioned API contracts for territorial data, evidence and product modules.",
    "title": "LoteDiretor Platform API",
    "version": "0.1.0"
  },
  "openapi": "3.0.0",
  "paths": {
    "/api/v1/health": {
      "get": {
        "operationId": "HealthController_health",
        "parameters": [],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "database": "ok",
                    "service": "platform-api",
                    "status": "ok",
                    "version": "0.1.0"
                  }
                }
              }
            },
            "description": ""
          }
        },
        "summary": "Platform API liveness/readiness status",
        "tags": [
          "health"
        ]
      }
    },
    "/api/v1/sources": {
      "get": {
        "operationId": "SourceRegistryController_list",
        "parameters": [
          {
            "in": "query",
            "name": "municipalityIbge",
            "required": false,
            "schema": {
              "example": "3550308",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SourceRegistryEntry"
                  },
                  "type": "array"
                }
              }
            },
            "description": "Source Registry entries"
          }
        },
        "summary": "List registered authoritative data sources",
        "tags": [
          "sources"
        ]
      }
    },
    "/api/v1/sources/{sourceCode}": {
      "get": {
        "operationId": "SourceRegistryController_get",
        "parameters": [
          {
            "in": "path",
            "name": "sourceCode",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "municipalityIbge",
            "required": false,
            "schema": {
              "example": "3550308",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourceRegistryEntry"
                }
              }
            },
            "description": "Source Registry entry"
          },
          "404": {
            "description": "Source not found"
          }
        },
        "summary": "Get one Source Registry entry",
        "tags": [
          "sources"
        ]
      }
    }
  },
  "servers": [],
  "tags": []
}